add database properties

+++ src/main/resources/db/migration/V1__create_tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-                       id SERIAL PRIMARY KEY,
+                       id BIGSERIAL PRIMARY KEY,
                        name VARCHAR(255) NOT NULL,
                        email VARCHAR(255) UNIQUE NOT NULL,
                        age INT CHECK (age >= 0 AND age <= 150),
@@ -10,7 +10,7 @@ CREATE TABLE users (
 );
 
 CREATE TABLE dishes (
-                        id SERIAL PRIMARY KEY,
+                        id BIGSERIAL PRIMARY KEY,
                         name VARCHAR(255) NOT NULL,
                         calories INT NOT NULL CHECK (calories >= 0),
                         proteins DOUBLE PRECISION NOT NULL CHECK (proteins >= 0),
@@ -19,15 +19,15 @@ CREATE TABLE dishes (
 );
 
 CREATE TABLE meals (
-                       id SERIAL PRIMARY KEY,
-                       user_id INT NOT NULL,
+                       id BIGSERIAL PRIMARY KEY,
+                       user_id BIGINT NOT NULL,
                        meal_date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
                        FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
 
 CREATE TABLE meal_dishes (
-                             meal_id INT NOT NULL,
-                             dish_id INT NOT NULL,
+                             meal_id BIGINT NOT NULL,
+                             dish_id BIGINT NOT NULL,
                              PRIMARY KEY (meal_id, dish_id),
                              FOREIGN KEY (meal_id) REFERENCES meals(id) ON DELETE CASCADE,
                              FOREIGN KEY (dish_id) REFERENCES dishes(id) ON DELETE CASCADE
